add new Release Podinfo workflow
This handles all of the push.tags events, and it anticipates that users
will not publish OCI manifests before CI has fully pushed an App Image.

Now "docker build+push" and "flux push artifact ...:production" are one
workflow and the deployment is not triggered if all prior steps did not
complete successfully.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>